patch: Update balenaOS helper
