+++ tsconfig.json
@@ -8,6 +8,7 @@
     "module": "commonjs",
     "target": "es2019",
     "outDir": "build",
+    "allowJs": true,
     "moduleResolution": "node",
     "allowSyntheticDefaultImports": true,
     "strict": true,
